backend/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
daphne==4.0.0
aioserial==1.3.1
requests==2.31.0
djangorestframework==3.15.2
